requirements.txt
opencv-python
pandas
matplotlib
ultralytics
pyyaml